+++ changelog.md
@@ -1,5 +1,15 @@
 # OceanWP Changelog
 
+### *2020.08.XX* - 1.8.7
+* **Added** - Hook: ocean_after_archive_product_price;
+* **Added** - WooCommerce: Product archive show/hide price and Add to Cart button option.
+* **Added** - WooCommerce: Product archive show/hide price and Add to Cart button message display (optional).
+* **Added** - WooCommerce: Product archive disable/enable image and title links option.
+* **Added** - WooCommerce: Product archive disable/enable image and title links conditional logic.
+* **Added** - WooCommerce: Single product show/hide price and Add to Cart button option.
+* **Added** - WooCommerce: Single product show/hode price and Add to Cart button message display (optional).
+* **Tweak** - WooCommerce: Product archive rating separated from price as an individual element.
+
 ### *2020.07.09* - 1.8.6:
 * **Fixed** - WooCommerce product gallery slider style image issue.
 * **Fixed** - Single blog post navigation text issue.